style: renamed shortinette package to tester for clarity
